split mongodb.js
