+++ lib/routes/webhook.js
@@ -14,8 +14,6 @@ var _receivedPostback2 = _interopRequireDefault(_receivedPostback);
 
 var _receivedMessage = require('../utils/receivedMessage');
 
-var _receivedMessage2 = _interopRequireDefault(_receivedMessage);
-
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 var VERIFY_TOKEN = _config2.default.VERIFY_TOKEN;
@@ -31,12 +29,12 @@ webhookRouter.get('/webhook/', function (req, res) {
 
 webhookRouter.post('/webhook', function (req, res) {
   var data = req.body;
-  // Make sure this is a page subscription
+
   if (data.object === 'page') {
     data.entry.forEach(function (pageEntry) {
       pageEntry.messaging.forEach(function (messagingEvent) {
         if (messagingEvent.message) {
-          (0, _receivedMessage2.default)(messagingEvent);
+          (0, _receivedMessage.receivedMessage)(messagingEvent);
         } else if (messagingEvent.postback) {
           // first time
           (0, _receivedPostback2.default)(messagingEvent);
