+++ lib/utils/receivedMessage.js
@@ -3,6 +3,7 @@
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
+exports.sendGenericMessageByArr = exports.receivedMessage = undefined;
 
 var _sendGenericMessage = require('./sendGenericMessage');
 
@@ -11,6 +12,26 @@ var _sendGenericMessage2 = _interopRequireDefault(_sendGenericMessage);
 var _sendTextMessage = require('./sendTextMessage');
 
 var _sendTextMessage2 = _interopRequireDefault(_sendTextMessage);
+
+var _findThreeVideos = require('../models/findThreeVideos');
+
+var _findThreeVideos2 = _interopRequireDefault(_findThreeVideos);
+
+var _saveLogData = require('../models/saveLogData');
+
+var _saveLogData2 = _interopRequireDefault(_saveLogData);
+
+var _saveSubscribeData = require('../models/saveSubscribeData');
+
+var _saveSubscribeData2 = _interopRequireDefault(_saveSubscribeData);
+
+var _removeSubscribeId = require('../models/removeSubscribeId');
+
+var _removeSubscribeId2 = _interopRequireDefault(_removeSubscribeId);
+
+var _findVideo = require('../models/findVideo');
+
+var _findVideo2 = _interopRequireDefault(_findVideo);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
@@ -38,32 +59,32 @@ var receivedMessage = function receivedMessage(event) {
   console.log('Received message for user ' + senderID + ' and page ' + recipientID + ' at ' + timeOfMessage + ' with message:');
 
   if (messageText === 'PPAV' || messageText === 'ppav' || messageText === 'Ppav') {
-    findThreeVideos().then(function (returnArr) {
+    (0, _findThreeVideos2.default)().then(function (returnArr) {
       sendGenericMessageByArr(senderID, returnArr);
-      saveLogData(true, {
+      (0, _saveLogData2.default)(true, {
         senderID: senderID,
         messageText: messageText,
         result: 'PPAV'
       });
     });
   } else if (messageText === 'GGinin' || messageText === 'GGININ' || messageText === 'gginin') {
-    saveSubscribeData(senderID).then(function (str) {
+    (0, _saveSubscribeData2.default)(senderID).then(function (str) {
       (0, _sendTextMessage2.default)(senderID, str);
     });
-  } else if (messageText === 'NoGG' || messageText === 'NOGG' || messageText === 'nogg' || messageText === 'noGG') {
-    removeSubscribeId(senderID).then(function (str) {
+  } else if (messageText === 'NoGG' || messageText === 'NOGG' || messageText === 'nogg' || messageText === 'noGG' || messageText === 'Nogg') {
+    (0, _removeSubscribeId2.default)(senderID).then(function (str) {
       (0, _sendTextMessage2.default)(senderID, str);
     });
   } else {
     switch (firstStr) {
       case '＃':
       case '#':
-        findVideo('code', messageText.split(firstStr)[1].toUpperCase()).then(function (returnObj) {
+        (0, _findVideo2.default)('code', messageText.split(firstStr)[1].toUpperCase()).then(function (returnObj) {
           var str = '';
           if (returnObj.results.length === 0) {
             str = '搜尋不到此番號';
             (0, _sendTextMessage2.default)(senderID, str);
-            saveLogData(false, {
+            (0, _saveLogData2.default)(false, {
               senderID: senderID,
               messageText: messageText,
               result: str
@@ -72,7 +93,7 @@ var receivedMessage = function receivedMessage(event) {
             str = '\u5E6B\u4F60\u641C\u5C0B\uFF1A' + returnObj.search_value;
             (0, _sendTextMessage2.default)(senderID, str);
             sendGenericMessageByArr(senderID, returnObj.results);
-            saveLogData(true, {
+            (0, _saveLogData2.default)(true, {
               senderID: senderID,
               messageText: messageText,
               result: str
@@ -82,12 +103,12 @@ var receivedMessage = function receivedMessage(event) {
         break;
       case '％':
       case '%':
-        findVideo('models', messageText.split(firstStr)[1]).then(function (returnObj) {
+        (0, _findVideo2.default)('models', messageText.split(firstStr)[1]).then(function (returnObj) {
           var str = '';
           if (returnObj.results.length === 0) {
             str = '搜尋不到此女優';
             (0, _sendTextMessage2.default)(senderID, str);
-            saveLogData(false, {
+            (0, _saveLogData2.default)(false, {
               senderID: senderID,
               messageText: messageText,
               result: str
@@ -96,7 +117,7 @@ var receivedMessage = function receivedMessage(event) {
             str = '\u5E6B\u4F60\u641C\u5C0B\uFF1A' + returnObj.search_value;
             (0, _sendTextMessage2.default)(senderID, str);
             sendGenericMessageByArr(senderID, returnObj.results);
-            saveLogData(true, {
+            (0, _saveLogData2.default)(true, {
               senderID: senderID,
               messageText: messageText,
               result: str
@@ -106,12 +127,12 @@ var receivedMessage = function receivedMessage(event) {
         break;
       case '＠':
       case '@':
-        findVideo('title', messageText.split(firstStr)[1]).then(function (returnObj) {
+        (0, _findVideo2.default)('title', messageText.split(firstStr)[1]).then(function (returnObj) {
           var str = '';
           if (returnObj.results.length === 0) {
             str = '搜尋不到此片名';
             (0, _sendTextMessage2.default)(senderID, str);
-            saveLogData(false, {
+            (0, _saveLogData2.default)(false, {
               senderID: senderID,
               messageText: messageText,
               result: str
@@ -120,7 +141,7 @@ var receivedMessage = function receivedMessage(event) {
             str = '\u5E6B\u4F60\u641C\u5C0B\uFF1A' + returnObj.search_value;
             (0, _sendTextMessage2.default)(senderID, str);
             sendGenericMessageByArr(senderID, returnObj.results);
-            saveLogData(true, {
+            (0, _saveLogData2.default)(true, {
               senderID: senderID,
               messageText: messageText,
               result: str
@@ -136,4 +157,5 @@ var receivedMessage = function receivedMessage(event) {
   }
 };
 
-exports.default = receivedMessage;
+exports.receivedMessage = receivedMessage;
+exports.sendGenericMessageByArr = sendGenericMessageByArr;
