+++ src/utils/updatePost.js
@@ -1,5 +1,5 @@
 import { findThreeNewVideos, findSubscribeId } from '../mongodb';
-import { sendGenericMessageByArr } from '../handleActions';
+import { sendGenericMessageByArr } from './receivedMessage';
 
 findThreeNewVideos().then(returnArr => {
   findSubscribeId().then(senderIDArr => {
